Massive structural change. Multipe controllers added. Two new layouts added.
