+++ WebShop/Controllers/HomeController.cs
@@ -19,135 +19,6 @@ namespace WebShop.Controllers
             return View();
         }
 
-        public ActionResult About_1()
-        {
-            return View();
-        }
-
-        public ActionResult About_2()
-        {
-            return View();
-        }
-
-        public ActionResult Authorisation()
-        {
-            return View();
-        }
-        public ActionResult Basket_step_1()
-        {
-            return View();
-        }
-
-        public ActionResult Basket_step_2()
-        {
-            return View();
-        }
-
-        public ActionResult Basket_step_3()
-        {
-            return View();
-        }
-
-        public ActionResult Basket_step_4()
-        {
-            return View();
-        }
-
-        public ActionResult Card()
-        {
-            return View();
-        }
-
-        public ActionResult Catalog_brands()
-        {
-            return View();
-        }
-
-        public ActionResult Empty_basket()
-        {
-            return View();
-        }
-
-        public ActionResult Error_404()
-        {
-            return View();
-        }
-
-        public ActionResult Error_500()
-        {
-            return View();
-        }
-
-        public ActionResult Faq()
-        {
-            return View();
-        }
-
-        public ActionResult Match_vincode()
-        {
-            return View();
-        }
-
-        public ActionResult News()
-        {
-            return View();
-        }
-
-        public ActionResult Orders()
-        {
-            return View();
-        }
-
-        public ActionResult Payment()
-        {
-            return View();
-        }
-
-        public ActionResult Profile_user()
-        {
-            return View();
-        }
-
-        public ActionResult Registration()
-        {
-            return View();
-        }
-
-        public ActionResult Serach_result()
-        {
-            return View();
-        }
-
-        public ActionResult Select_car()
-        {
-            return View();
-        }
-
-        public ActionResult Sto()
-        {
-            return View();
-        }
-
-        public ActionResult Upload()
-        {
-            return View();
-        }
-        public ActionResult Clients()
-        {
-            return View();
-        }
-        public ActionResult Admin_pan()
-        {
-            return View();
-        }
-        public ActionResult Catalog()
-        {
-            return View();
-        }
-        public ActionResult Vacation()
-        {
-            return View();
-        }
         [HttpPost]
         public ActionResult Do ()
         {
