+++ WebShop/Controllers/InformationalController.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+
+namespace WebShop.Controllers
+{
+    public class InformationalController : Controller
+    {
+        // GET: Informational
+        public ActionResult About_1()
+        {
+            return View();
+        }
+
+        public ActionResult About_2()
+        {
+            return View();
+        }
+
+        public ActionResult Faq()
+        {
+            return View();
+        }
+
+        public ActionResult News()
+        {
+            return View();
+        }
+
+        public ActionResult Vacation()
+        {
+            return View();
+        }
+    }
+}
