+++ WebShop/Views/Admin/Clients.cshtml
@@ -1,0 +1,12 @@
+﻿
+@{
+    Layout = "~/Views/Shared/_PersonalLayout.cshtml";
+    ViewBag.Title = "All users";
+}
+
+<main class="main-content">
+    <div class="main-heading">
+        <div class="main-heading__text">Клиенты</div>
+    </div>
+    <h2>Тут будет выводится список всех пользователей</h2>
+</main>
