+++ WebShop/Views/BasketPayment/Basket_step_2.cshtml
@@ -1,0 +1,97 @@
+﻿
+@{
+    Layout = "~/Views/Shared/_PersonalLayout.cshtml";
+    ViewBag.Title = "Basket_step_2";
+}
+
+<main class="main-content">
+    <h1 class="main-title">Заказ<span>№ 00000154705@19</span></h1>
+    <div class="order-steps-wrap text-center">
+        <ul class="order-steps">
+            <li class="order-steps__item">
+                <svg class="icon icon-shopping-cart order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#shopping-cart"></use>
+                </svg><span class="order-steps__title">Корзина</span>
+            </li>
+            <li class="order-steps__item is-current">
+                <svg class="icon icon-wallet-strong order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#wallet-strong"></use>
+                </svg><span class="order-steps__title">Оплата</span>
+            </li>
+            <li class="order-steps__item">
+                <svg class="icon icon-lorry order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#lorry"></use>
+                </svg><span class="order-steps__title">Доставка</span>
+            </li>
+            <li class="order-steps__item">
+                <svg class="icon icon-confirmation order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#confirmation"></use>
+                </svg><span class="order-steps__title">Завершение</span>
+            </li>
+        </ul>
+    </div>
+    <div class="payment-section">
+        <div class="payment-section__heading">Форма оплаты</div>
+        <div class="radio-option">
+            <div class="row">
+                <div class="col-sm-6">
+                    <div class="radio-opt active">
+                        <div class="radio-opt__title">Оплата онлайн</div>
+                        <div class="radio-opt__user-data">
+                            <div class="row">
+                                <div class="col-lg-8">
+                                </div>
+                                <div class="col-lg-4">
+                                </div>
+                            </div>
+                        </div>
+                        <div class="radio-opt__footer">
+                            <div class="radio-label">
+                                <input class="radio-label__input" type="radio" name="payment" value="payment-online" checked id="opt-1">
+                                <label class="radio-label__main" for="opt-1">Выбрать этот способ оплаты</label>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="col-sm-6">
+                    <div class="radio-opt">
+                        <div class="radio-opt__title">Оплата наличными</div>
+                        <figure class="radio-opt__figure">
+                            <svg class="icon radio-opt__icon" viewbox="0 0 512 512">
+                                <use href="~/assets/images/svg/symbol/sprite.svg#money"></use>
+                            </svg>
+                            <figcaption class="radio-opt__figcaption"><strong>Оплата заказа наличными курьеру либо в кассу</strong></figcaption>
+                        </figure>
+                        <div class="radio-opt__footer">
+                            <div class="radio-label">
+                                <input class="radio-label__input" type="radio" name="payment" value="payment-cash" id="opt-2">
+                                <label class="radio-label__main" for="opt-2">Выбрать этот способ оплаты</label>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="order-footer d-none d-md-flex">
+        <a class="btn btn--reverse" href="@Url.Action("Basket_step_1", "BasketPayment")">Назад</a>
+        <textarea class="order-footer__message" placeholder="Комментарий к заказу" rows="1" name="order-message"></textarea>
+        <div class="footer-summary-price">
+            <div class="footer-summary-price__title">Сумма к оплате</div>
+            <div class="footer-summary-price__value">2 297 <span>Лей</span></div>
+        </div><a class="btn btn--main" href="@Url.Action("Basket_step_3", "BasketPayment")">Далее<i class="fas fa-chevron-circle-right"></i></a>
+    </div>
+    <div class="order-mob-footer d-md-none">
+        <footer class="order-mob-footer__main">
+            <p class="order-mob-footer__payment">Сумма к оплате</p>
+            <div class="order-mob-footer__total-price"><span class="order-mob-footer__total-price-value">2 297</span> Лей</div>
+            <textarea class="order-mob-footer__message" placeholder="Комментарий к заказу" rows="2" name="order-message"></textarea>
+        </footer>
+        <div class="order-mob-footer__controls">
+            <div class="row">
+                <div class="col-4"><a class="btn btn--reverse" href="@Url.Action("Basket_step_1", "BasketPayment")">Назад</a></div>
+                <div class="col-8"><a class="btn btn--main" href="@Url.Action("Basket_step_3", "BasketPayment")">Далее<i class="fas fa-chevron-circle-right"></i></a></div>
+            </div>
+        </div>
+    </div>
+</main>
