+++ WebShop/Views/BasketPayment/Basket_step_4.cshtml
@@ -1,0 +1,72 @@
+﻿
+@{
+    Layout = "~/Views/Shared/_PersonalLayout.cshtml";
+    ViewBag.Title = "Basket_step_4";
+}
+
+<main class="main-content">
+    <h1 class="main-title">Заказ<span>№ 00000154705@19</span></h1>
+    <div class="order-steps-wrap text-center">
+        <ul class="order-steps">
+            <li class="order-steps__item">
+                <svg class="icon icon-shopping-cart order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#shopping-cart"></use>
+                </svg><span class="order-steps__title">Корзина</span>
+            </li>
+            <li class="order-steps__item">
+                <svg class="icon icon-wallet-strong order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#wallet-strong"></use>
+                </svg><span class="order-steps__title">Оплата</span>
+            </li>
+            <li class="order-steps__item">
+                <svg class="icon icon-lorry order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#lorry"></use>
+                </svg><span class="order-steps__title">Доставка</span>
+            </li>
+            <li class="order-steps__item is-current">
+                <svg class="icon icon-confirmation order-steps__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#confirmation"></use>
+                </svg><span class="order-steps__title">Завершение</span>
+            </li>
+        </ul>
+    </div>
+    <figure class="order-success">
+        <img class="order-success__img img-responsive" src="~/assets/images/general/order-success.svg" alt="Order Success" width="90" height="90">
+        <figcaption class="order-success__figcaption">
+            <div class="order-success__heading">Благодарим за покупку</div>
+            <p class="order-success__desc">Ваш заказ <a href="#">№00000154705@19</a> был принят нами, и в ближайшее время Вы его получите.</p>
+            <p class="order-success__desc">Если у Вас возникли вопросы по заказу, обратитесь в отдел продаж:</p><a class="tiny-contacts__wrap" href="tel:7333">
+                <svg class="icon icon-phone-call tiny-contacts__icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#phone-call"></use>
+                </svg><span class="tiny-contacts__phone">+373 79-637-411</span>
+            </a>
+        </figcaption>
+    </figure>
+    <div class="manual-params manual-params--bigger">
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Склад отгрузки</dt>
+            <dd class="manual-params__property">Кишинев, Studenților 9/9</dd>
+        </dl>
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Заказ доступен с</dt>
+            <dd class="manual-params__property">06 Ноября 2025 / 14:00</dd>
+        </dl>
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Форма оплаты</dt>
+            <dd class="manual-params__property">Безналичный</dd>
+        </dl>
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Тип доставки</dt>
+            <dd class="manual-params__property">Доставка автотранспортом</dd>
+        </dl>
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Адрес доставки</dt>
+            <dd class="manual-params__property">Republica Moldova, or.Codru, mun.Chișinău, ул.Qwerty, дом № 94/10</dd>
+        </dl>
+        <dl class="manual-params__param">
+            <dt class="manual-params__value">Комментарий</dt>
+            <dd class="manual-params__property">Привезите пж в офис</dd>
+        </dl>
+    </div>
+    <div class="basket-success-summary"><a class="btn btn--main" href="@Url.Action("MainPage", "Home")">Вернуться на главную</a></div>
+</main>
