+++ WebShop/Views/Error/Error_500.cshtml
@@ -12,7 +12,7 @@
         <figure class="error-notification">
             <div class="error-notification__img-wrap"><img class="error-notification__img img-responsive" src="~/assets/images/general/500.png" alt="Страница не найдена"></div>
             <figcaption class="error-notification__figcaption">
-                <div class="error-notification__desc">Так уж получилось, что из множества страниц нашего сайта Вы оказались как раз на той, которая уже не существует… Предлагаем вам вернуться на главную страницу.</div><a class="btn btn--main" href="#">Вернуться на главную</a>
+                <div class="error-notification__desc">Так уж получилось, что из множества страниц нашего сайта Вы оказались как раз на той, которая уже не существует… Предлагаем вам вернуться на главную страницу.</div><a class="btn btn--main" href="@Url.Action("MainPage", "Home")">Вернуться на главную</a>
             </figcaption>
         </figure>
     </div>
