--- WebShop/Views/Home/About_1.cshtml
@@ -1,64 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "About_1";
-}
-
-<div class="container-fluid">
-    <div class="primary-layout">
-        <h1 class="main-title">О компании</h1>
-        <div class="primary-content">
-            <div class="sidebar">
-                <div class="category-nav d-none d-md-block">
-                    <ul class="category-nav__list">
-                        <li class="category-nav__item"><a class="category-nav__link important" href="#">Акции</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">О нас</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">История</a></li>
-                        <li class="category-nav__item active"><a class="category-nav__link" href="#">Миссия и философия</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Цели</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Обязательства</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Руководство</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Сотрудничество</a></li>
-                    </ul>
-                </div>
-            </div>
-            <main class="main-content">
-                <div class="post">
-                    <h2>Миссия и философия AutoWares</h2>
-                    <p><img class="img-responsive" src="~/assets/images/content/post-img.jpg" alt="AutoWares"></p>
-                    <p><strong>Компания AutoWares</strong>  является представителем ведущих Европейских заводов производителей.</p>
-                    <p>AutoWares - оптовая компания по продажам - автозапчастей, моторных масел, смазочных материалов, автохимии, автоаксессуаров, автокосметики.</p>
-                    <p>Основные направления компании это дилеры, оптовые, розничные продажи по Казахстану.</p>
-                    <p>Каждый клиент, работая с нами, получает персонального менеджера, который отслеживает всю цепочку от заказа до его получения. Быстрый документооборот и любая форма оплаты.</p>
-                    <h3 class="text-center">Миссия AutoWares – делать все, чтобы автомобили оставались в пути.</h3>
-                    <div class="row">
-                        <div class="col-md-6">
-                            <strong>Мы предлагаем Вам:</strong>
-                            <ol>
-                                <li>Широкий ассортимент автозапчастей.</li>
-                                <li>Высокий процент наличия на складе.</li>
-                                <li>Гибкая система скидок для постоянных клиентов.</li>
-                                <li>Срочную доставку автозапчастей по городу Кишиневу.</li>
-                                <li>Квалифицированную помощь в подборе запчастей.</li>
-                                <li>Проведение обучающих семинаров.</li>
-                                <li>Предоставление рекламной продукции.</li>
-                            </ol>
-                        </div>
-                        <div class="col-md-6">
-                            <strong>Это позволит Вам:</strong>
-                            <ol>
-                                <li>Предоставлять дополнительный сервис для Ваших клиентов.</li>
-                                <li>Сократить время на поиск и доставку автозапчастей.</li>
-                                <li>Сократить время простоя обслуживающегося автомобиля.</li>
-                                <li>Развивать Ваш бизнес в новом направлении.</li>
-                                <li>Быть уверенным в качестве автозапчастей.</li>
-                                <li>Быть конкурентными по цене и при этом получать дополнительную прибыль.</li>
-                                <li>Получить квалифицированную помощь и профессиональные знания.</li>
-                            </ol>
-                        </div>
-                    </div>
-                </div>
-            </main>
-        </div>
-    </div>
-</div>
-
