--- WebShop/Views/Home/Empty_basket.cshtml
@@ -1,66 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "Empty_basket";
-}
-
-<div class="container-fluid">
-    <div class="primary-layout">
-        <div class="primary-content">
-            <div class="sidebar">
-                <div class="category-nav d-none d-md-block">
-                    <ul class="category-nav__list">
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Корзина</a></li>
-                        <li class="category-nav__item active"><a class="category-nav__link" href="#">Заказы</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Оплата</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Акт сверки</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Клиенты</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Профиль</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">VIN запросы</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Мероприятия</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Акции</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Управление пользователями</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Регистрация пользователей</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Статистика доступа</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Контент групп</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Управление сайтом</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Загрузка прайса/кроссов</a></li>
-                    </ul>
-                </div>
-            </div>
-            <main class="main-content">
-                <h1 class="main-title">Корзина</h1>
-                <div class="order-steps-wrap text-center">
-                    <ul class="order-steps">
-                        <li class="order-steps__item is-current">
-                            <svg class="icon icon-shopping-cart order-steps__icon">
-                                <use href="~/assets/images/svg/symbol/sprite.svg#shopping-cart"></use>
-                            </svg><span class="order-steps__title">Корзина</span>
-                        </li>
-                        <li class="order-steps__item">
-                            <svg class="icon icon-wallet-strong order-steps__icon">
-                                <use href="~/assets/images/svg/symbol/sprite.svg#wallet-strong"></use>
-                            </svg><span class="order-steps__title">Оплата</span>
-                        </li>
-                        <li class="order-steps__item">
-                            <svg class="icon icon-lorry order-steps__icon">
-                                <use href="~/assets/images/svg/symbol/sprite.svg#lorry"></use>
-                            </svg><span class="order-steps__title">Доставка</span>
-                        </li>
-                        <li class="order-steps__item">
-                            <svg class="icon icon-confirmation order-steps__icon">
-                                <use href="~/assets/images/svg/symbol/sprite.svg#confirmation"></use>
-                            </svg><span class="order-steps__title">Завершение</span>
-                        </li>
-                    </ul>
-                </div>
-                <div class="empty-basket text-center">
-                    <header class="orders-list__header">
-                        <div class="orders-list__heading">Ваша корзина пуста</div>
-                    </header>
-                    <div class="empty-basket__img-wrap"><img class="empty-basket__img img-responsive" src="~/assets/images/general/basket-default.png" alt=""></div><a class="empty-basket__btn btn btn--main" href="#">Перейти на главную</a>
-                </div>
-            </main>
-        </div>
-    </div>
-</div>
-
