--- WebShop/Views/Home/Payment.cshtml
@@ -1,58 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "Payment";
-}
-
-<div class="container-fluid">
-    <ul class="breadcrumbs">
-        <li class="breadcrumbs__item"><a class="breadcrumbs__link" href="#">Главная</a></li>
-        <li class="breadcrumbs__item"><a class="breadcrumbs__link" href="#">Автохимия</a></li>
-        <li class="breadcrumbs__item"><a class="breadcrumbs__link" href="#">Система топливная</a></li>
-        <li class="breadcrumbs__item">Присадки к бензину</li>
-    </ul>
-    <div class="primary-layout">
-        <div class="primary-content">
-            <div class="sidebar">
-                <div class="category-nav d-none d-md-block">
-                    <ul class="category-nav__list">
-                        <li class="category-nav__item active"><a class="category-nav__link" href="#">Каталоги производителей</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Аккумуляторы</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Газовые пружины</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Кузов/Оптика</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Мотор и сцепление</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">OEM Каталоги</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Система охлаждения</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Система очистки окон</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Тормозная система</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Фильтры и масла</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Шины и диски</a></li>
-                        <li class="category-nav__item"><a class="category-nav__link" href="#">Электрика</a></li>
-                    </ul>
-                </div>
-            </div>
-            <main class="main-content">
-                <div class="main-heading">
-                    <h1 class="main-heading__text">Оплата</h1>
-                    <div class="main-heading__link-wrap"> <strong>текущая задолженность<span>0 Лей</span></strong></div>
-                </div>
-                <form class="payment-form" action="/" method="POST" name="payment-form">
-                    <div class="payment-form__fields">
-                        <div class="payment-form__field">
-                            <div class="payment-form__helper">Введите данные для оплаты</div>
-                        </div>
-                        <div class="payment-form__field">
-                            <input class="payment-form__input" type="email" name="user-email" placeholder="Email плательщика">
-                        </div>
-                        <div class="payment-form__field payment-form__field--currency">
-                            <input class="payment-form__input" type="text" name="user-price" placeholder="Сумма к оплате"><span class="payment-form__currency">Лей</span>
-                        </div>
-                        <div class="payment-form__field">
-                            <input class="payment-form__btn btn btn--main" type="submit" name="send-data" value="Оплатить">
-                        </div>
-                    </div>
-                </form>
-            </main>
-        </div>
-    </div>
-</div>
-
