+++ WebShop/Views/Informational/About_1.cshtml
@@ -1,0 +1,43 @@
+﻿
+@{
+    Layout = "~/Views/Shared/_InfoLayout.cshtml";
+    ViewBag.Title = "About_1";
+}
+
+<main class="main-content">
+    <div class="post">
+        <h2>Миссия и философия AutoWares</h2>
+        <p><img class="img-responsive" src="~/assets/images/content/post-img.jpg" alt="AutoWares"></p>
+        <p><strong>Компания AutoWares</strong>  является представителем ведущих Европейских заводов производителей.</p>
+        <p>AutoWares - оптовая компания по продажам - автозапчастей, моторных масел, смазочных материалов, автохимии, автоаксессуаров, автокосметики.</p>
+        <p>Основные направления компании это дилеры, оптовые, розничные продажи по Молдова.</p>
+        <p>Каждый клиент, работая с нами, получает персонального менеджера, который отслеживает всю цепочку от заказа до его получения. Быстрый документооборот и любая форма оплаты.</p>
+        <h3 class="text-center">Миссия AutoWares – делать все, чтобы автомобили оставались в пути.</h3>
+        <div class="row">
+            <div class="col-md-6">
+                <strong>Мы предлагаем Вам:</strong>
+                <ol>
+                    <li>Широкий ассортимент автозапчастей.</li>
+                    <li>Высокий процент наличия на складе.</li>
+                    <li>Гибкая система скидок для постоянных клиентов.</li>
+                    <li>Срочную доставку автозапчастей по городу Кишиневу.</li>
+                    <li>Квалифицированную помощь в подборе запчастей.</li>
+                    <li>Проведение обучающих семинаров.</li>
+                    <li>Предоставление рекламной продукции.</li>
+                </ol>
+            </div>
+            <div class="col-md-6">
+                <strong>Это позволит Вам:</strong>
+                <ol>
+                    <li>Предоставлять дополнительный сервис для Ваших клиентов.</li>
+                    <li>Сократить время на поиск и доставку автозапчастей.</li>
+                    <li>Сократить время простоя обслуживающегося автомобиля.</li>
+                    <li>Развивать Ваш бизнес в новом направлении.</li>
+                    <li>Быть уверенным в качестве автозапчастей.</li>
+                    <li>Быть конкурентными по цене и при этом получать дополнительную прибыль.</li>
+                    <li>Получить квалифицированную помощь и профессиональные знания.</li>
+                </ol>
+            </div>
+        </div>
+    </div>
+</main>
