+++ WebShop/Views/Profile/Profile_user.cshtml
@@ -1,0 +1,85 @@
+﻿
+@{
+    Layout = "~/Views/Shared/_PersonalLayout.cshtml";
+    ViewBag.Title = "Profile";
+}
+
+<main class="main-content">
+    <h1 class="main-title">Профиль</h1>
+    <div class="section-area">
+        <div class="section-area__heading">Личные данные</div>
+        <div class="manual-params manual-params--bigger">
+            <dl class="manual-params__param">
+                <dt class="manual-params__value">Ваша Фамилия</dt>
+                <dd class="manual-params__property">Иванов</dd>
+            </dl>
+            <dl class="manual-params__param">
+                <dt class="manual-params__value">Ваше Имя</dt>
+                <dd class="manual-params__property">Николай</dd>
+            </dl>
+            <dl class="manual-params__param">
+                <dt class="manual-params__value">Email (используется как логин)</dt>
+                <dd class="manual-params__property">root@webmall.md</dd>
+            </dl>
+            <dl class="manual-params__param">
+                <dt class="manual-params__value">Телефон мобильный</dt>
+                <dd class="manual-params__property">7 037312312</dd>
+            </dl>
+            <dl class="manual-params__param">
+                <dt class="manual-params__value">Телефон стационарный</dt>
+                <dd class="manual-params__property">7 123</dd>
+            </dl><a class="manual-params__edit" href="#">
+                <svg class="icon icon-pen manual-params__edit-icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#pen"></use>
+                </svg>Редактировать
+            </a>
+        </div>
+    </div>
+    <form action="/" method="POST" name="profile-form">
+        <div class="section-area">
+            <div class="section-area__heading">Изменить пароль</div>
+            <div class="row">
+                <div class="col-sm-4">
+                    <div class="section-area__field">
+                        <label class="section-area__label">Текущий пароль <span>*</span></label>
+                        <input class="section-area__input" type="password" name="current-password">
+                    </div>
+                </div>
+                <div class="col-sm-4">
+                    <div class="section-area__field">
+                        <label class="section-area__label">Новый пароль <span>*</span></label>
+                        <input class="section-area__input section-area__input--gray" type="password" name="new-password">
+                    </div>
+                </div>
+                <div class="col-sm-4">
+                    <div class="section-area__field">
+                        <label class="section-area__label">Новый пароль еще раз <span>*</span></label>
+                        <input class="section-area__input section-area__input--gray" type="password" name="new-password-check">
+                    </div>
+                </div>
+            </div><a class="section-area__save-psw" href="#">
+                <svg class="icon icon-save section-area__save-psw-icon">
+                    <use href="~/assets/images/svg/symbol/sprite.svg#save"></use>
+                </svg>Сохранить изменения пароля
+            </a>
+        </div>
+        <div class="section-area">
+            <div class="section-area__heading">Настройки и уведомления</div>
+            <div class="section-area__checkbox checkbox-label">
+                <input class="checkbox-label__input" type="checkbox" name="is-available" id="opt-1">
+                <label class="checkbox-label__main" for="opt-1">Розничная цена в карточке товара</label>
+            </div>
+            <div class="section-area__checkbox checkbox-label">
+                <input class="checkbox-label__input" type="checkbox" name="is-available" id="opt-2">
+                <label class="checkbox-label__main" for="opt-2">Присылать уведомление о заказе на e-mail</label>
+            </div>
+            <div class="section-area__checkbox checkbox-label">
+                <input class="checkbox-label__input" type="checkbox" name="is-available" id="opt-3">
+                <label class="checkbox-label__main" for="opt-3">Подписаться на информацию об акциях</label>
+            </div>
+            <div class="section-area__footer">
+                <input class="btn btn--main" type="submit" name="send-data" value="Сохранить изменения">
+            </div>
+        </div>
+    </form>
+</main>
