+++ WebShop/Views/Shared/_InfoLayout.cshtml
@@ -1,0 +1,38 @@
+﻿@{
+    Layout = "~/Views/Shared/_Layout.cshtml";
+    var isAdmin = true;
+}
+
+@section Styles {
+    @RenderSection("Styles", false)
+}
+<div class="container-fluid">
+
+    <div class="primary-layout">
+        <div class="primary-content">
+            <div class="sidebar">
+                <div class="category-nav d-none d-md-block">
+                    <ul class="category-nav__list">
+                        <li class="category-nav__item"><a class="category-nav__link @*important*@" href="@Url.Action("Search_result", "CatalogSearch")">Акции</a></li>
+                        <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("About_2", "Informational")">О нас</a></li>
+                        <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("About_1", "Informational")">Миссия и философия</a></li>
+                        <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("News", "Informational")">Новости</a></li>
+                        <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("Vacation", "Informational")">Вакансии</a></li>
+                        <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("Faq", "Informational")">Часто задаваемые вопросы</a></li>
+                        @if (isAdmin)
+                        {
+                            <li class="category-nav__item"><a class="category-nav__link" href="@Url.Action("Clients", "Admin")">Управление пользователями</a></li>
+                            <li class="category-nav__item"><a class="category-nav__link" href="#">Регистрация пользователей</a></li>
+                            <li class="category-nav__item"><a class="category-nav__link" href="#">Управление сайтом</a></li>
+                        }
+                    </ul>
+                </div>
+            </div>
+            @RenderBody()
+        </div>
+    </div>
+</div>
+
+@section Scripts {
+    @RenderSection("Scripts", required: false)
+}
