+++ WebShop/Views/Shared/_Layout.cshtml
@@ -7,11 +7,12 @@
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <meta name="keywords" content="">
     <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&amp;display=swap" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="~/assets/css/styles.min.css">
+    @*<link rel="stylesheet" type="text/css" href="~/assets/css/styles.min.css">*@
     @*<link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" asp-append-version="true" />*@
-    @Styles.Render("~/bundles/bootstrap/css")
+    @*@Styles.Render("~/bundles/bootstrap/css")*@
+    @Styles.Render("~/assets/css/all")
+    @RenderSection("Styles", false)
 
-    
 </head>
 <body class="index-page">
 
@@ -22,13 +23,15 @@
 
     @Html.Partial("_Footer")
 
-    <script src="~/assets/js/libs.min.js"></script>
+    @*<script src="~/assets/js/libs.min.js"></script>*@
     <script src="https://unpkg.com/popper.js@1"></script>
     <script src="https://unpkg.com/tippy.js@5"></script>
     @*<script src="~/assets/js/Modal.js"></script>*@
-    <script src="~/assets/js/main.js"></script>
+    @*<script src="~/assets/js/main.js"></script>*@
     @*<script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js" asp-append-version="true"></script>*@
-    @Scripts.Render("~/bundles/bootstrap/js")
+    @*@Scripts.Render("~/bundles/bootstrap/js")*@
     @Scripts.Render("~/bundles/jquery")
+    @Scripts.Render("~/assets/js/all")
+    @RenderSection("Scripts", required: false)
 </body>
 </html>
