+++ WebShop/WebShop.csproj
@@ -302,8 +302,18 @@
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="Controllers\AdminController.cs" />
+    <Compile Include="Controllers\AuthController.cs" />
+    <Compile Include="Controllers\BasketPaymentController.cs" />
+    <Compile Include="Controllers\CatalogSearchController.cs" />
+    <Compile Include="Controllers\ErrorController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
+    <Compile Include="Controllers\InformationalController.cs" />
+    <Compile Include="Controllers\OtherServicesController.cs" />
+    <Compile Include="Controllers\ProductController.cs" />
+    <Compile Include="Controllers\ProfileController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
@@ -330,36 +340,37 @@
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\_ViewStart.cshtml" />
     <Content Include="Views\Home\MainPage.cshtml" />
-    <Content Include="Views\Home\Upload.cshtml" />
-    <Content Include="Views\Home\Payment.cshtml" />
-    <Content Include="Views\Home\Basket_step_4.cshtml" />
-    <Content Include="Views\Home\Basket_step_3.cshtml" />
-    <Content Include="Views\Home\Basket_step_2.cshtml" />
-    <Content Include="Views\Home\Basket_step_1.cshtml" />
-    <Content Include="Views\Home\Orders.cshtml" />
-    <Content Include="Views\Home\Match_vincode.cshtml" />
-    <Content Include="Views\Home\Profile_user.cshtml" />
-    <Content Include="Views\Home\Select_car.cshtml" />
-    <Content Include="Views\Home\Sto.cshtml" />
-    <Content Include="Views\Home\Error_404.cshtml" />
-    <Content Include="Views\Home\Error_500.cshtml" />
-    <Content Include="Views\Home\About_1.cshtml" />
-    <Content Include="Views\Home\About_2.cshtml" />
-    <Content Include="Views\Home\Empty_basket.cshtml" />
-    <Content Include="Views\Home\Authorisation.cshtml" />
-    <Content Include="Views\Home\Vacation.cshtml" />
-    <Content Include="Views\Home\Card.cshtml" />
-    <Content Include="Views\Home\Catalog_brands.cshtml" />
-    <Content Include="Views\Home\News.cshtml" />
-    <Content Include="Views\Home\Registration.cshtml" />
-    <Content Include="Views\Home\Search_result.cshtml" />
-    <Content Include="Views\Home\Faq.cshtml" />
+    <Content Include="Views\BasketPayment\Payment.cshtml" />
+    <Content Include="Views\BasketPayment\Basket_step_4.cshtml" />
+    <Content Include="Views\BasketPayment\Basket_step_3.cshtml" />
+    <Content Include="Views\BasketPayment\Basket_step_2.cshtml" />
+    <Content Include="Views\Profile\Orders.cshtml" />
+    <Content Include="Views\CatalogSearch\Match_vincode.cshtml" />
+    <Content Include="Views\Profile\Profile_user.cshtml" />
+    <Content Include="Views\CatalogSearch\Select_car.cshtml" />
+    <Content Include="Views\OtherServices\Sto.cshtml" />
+    <Content Include="Views\Error\Error_404.cshtml" />
+    <Content Include="Views\Error\Error_500.cshtml" />
+    <Content Include="Views\Informational\About_1.cshtml" />
+    <Content Include="Views\Informational\About_2.cshtml" />
+    <Content Include="Views\BasketPayment\Empty_basket.cshtml" />
+    <Content Include="Views\Auth\Authorisation.cshtml" />
+    <Content Include="Views\Informational\Vacation.cshtml" />
+    <Content Include="Views\Product\Card.cshtml" />
+    <Content Include="Views\CatalogSearch\Catalog_brands.cshtml" />
+    <Content Include="Views\Informational\News.cshtml" />
+    <Content Include="Views\Auth\Registration.cshtml" />
+    <Content Include="Views\CatalogSearch\Search_result.cshtml" />
+    <Content Include="Views\Informational\Faq.cshtml" />
     <Content Include="Views\Shared\_Header.cshtml" />
     <Content Include="Views\Shared\_Footer.cshtml" />
-    <Content Include="Views\Home\Clients.cshtml" />
-    <Content Include="Views\Home\Admin_pan.cshtml" />
-    <Content Include="Views\Home\Catalog.cshtml" />
+    <Content Include="Views\Admin\Clients.cshtml" />
+    <Content Include="Views\Admin\Admin_pan.cshtml" />
+    <Content Include="Views\CatalogSearch\Catalog.cshtml" />
     <Content Include="Views\Shared\_Layout.cshtml" />
+    <Content Include="Views\Shared\_PersonalLayout.cshtml" />
+    <Content Include="Views\BasketPayment\Basket_step_1.cshtml" />
+    <Content Include="Views\Shared\_InfoLayout.cshtml" />
     <None Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
     </None>
@@ -370,6 +381,7 @@
   <ItemGroup>
     <Folder Include="App_Data\" />
     <Folder Include="Models\" />
+    <Folder Include="Scripts\" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
